+++ cypress/integration/core/toggletip.spec.ts
@@ -1,0 +1,43 @@
+describe('Toggletip', () => {
+  beforeEach(() => {
+    cy.visit('http://localhost:56106/iframe.html?id=toggletip--toggletip');
+  });
+
+  function prepareComponent() {
+    cy
+      .get('dso-toggletip')
+      .as('dsoToggletip');
+
+    cy
+      .get('@dsoToggletip')
+      .shadow()
+      .find('dso-info-button')
+      .as('dsoInfoButton');
+
+    cy
+      .get('@dsoInfoButton')
+      .shadow()
+      .find('button')
+      .as('dsoButton');
+
+    cy
+      .get('@dsoToggletip')
+      .shadow()
+      .find('dso-tooltip')
+      .as('dsoTooltip');
+  }
+
+  it('should focus on button after escape', () => {
+    prepareComponent();
+
+    cy
+      .get('@dsoButton')
+      .click()
+      .get('@dsoTooltip')
+      .should('not.have.class', 'hidden')
+      .realPress('Tab')
+      .realPress('Escape')
+      .get('@dsoButton')
+      .should('have.focus');
+  });
+});